requirements.txt
alembic==1.13.1
asttokens==2.4.1
black==24.1.0
blinker==1.7.0
click==8.1.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
flake8==7.0.0
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
ipython==8.20.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
